% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/introduce_myself.R
\name{introduce_myself}
\alias{introduce_myself}
\title{introduce myself}
\usage{
introduce_myself()
}
\value{
a string with an introduction about me
}
\description{
introduce myself
}
\examples{
introduce_myself()
}
