% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/introduce_yourself.R
\name{introduce_yourself}
\alias{introduce_yourself}
\title{introduce yourself}
\usage{
introduce_yourself(name)
}
\arguments{
\item{name}{your name to introduce}
}
\value{
a string with an introduction about you
}
\description{
introduce yourself
}
\examples{
introduce_yourself("Arthur")
}
